+++ packages/react/src/components/DateTimePicker/DateTimePickerV2.test.jsx
@@ -385,17 +385,18 @@ describe('DateTimePickerV2', () => {
     expect(dateTimePickerProps.onApply).toHaveBeenCalled();
   });
 
-  it('should render with a predefined absolute range (new time spinner)', () => {
+  it('should render with a predefined absolute range (new time spinner and 12 hours time)', () => {
     render(
       <DateTimePicker
         useNewTimeSpinner
+        dateTimeMask="YYYY-MM-DD hh:mm A"
         {...dateTimePickerProps}
         defaultValue={defaultAbsoluteValue}
       />
     );
 
     // default value starts at   '2020-04-01' at 12:34 to 2020-04-06 at 10:49
-    expect(screen.getByText('2020-04-01 12:34 to 2020-04-06 10:49')).toBeVisible();
+    expect(screen.getByText('2020-04-01 12:34 PM to 2020-04-06 10:49 AM')).toBeVisible();
 
     // first open the menu
     userEvent.click(screen.getAllByLabelText('Calendar')[0]);
@@ -437,7 +438,53 @@ describe('DateTimePickerV2', () => {
     expect(dateTimePickerProps.onApply).toHaveBeenCalled();
   });
 
-  it('should render with a predefined single select date and time', () => {
+  it('should render with a predefined absolute range (new time spinner and 24 hours format time)', () => {
+    render(
+      <DateTimePicker
+        useNewTimeSpinner
+        {...dateTimePickerProps}
+        defaultValue={defaultAbsoluteValue}
+      />
+    );
+
+    // default value starts at   '2020-04-01' at 12:34 to 2020-04-06 at 10:49
+    expect(screen.getByText('2020-04-01 12:34 to 2020-04-06 10:49')).toBeVisible();
+
+    // first open the menu
+    userEvent.click(screen.getAllByLabelText('Calendar')[0]);
+
+    // open start time picker
+    fireEvent.focus(screen.getByTestId('date-time-picker-input-1'));
+
+    // select hour
+    userEvent.click(
+      within(screen.getByTestId('date-time-picker-spinner-list-spinner-1')).getByText('13')
+    );
+    // select miniute
+    userEvent.click(
+      within(screen.getByTestId('date-time-picker-spinner-list-spinner-2')).getByText('34')
+    );
+
+    screen.getByTestId('date-time-picker-spinner').blur();
+
+    // open end time picker
+    fireEvent.focus(screen.getByTestId('date-time-picker-input-2'));
+    // select hour
+    userEvent.click(
+      within(screen.getByTestId('date-time-picker-spinner-list-spinner-1')).getByText('11')
+    );
+    // select minute
+    userEvent.click(
+      within(screen.getByTestId('date-time-picker-spinner-list-spinner-2')).getByText('49')
+    );
+
+    screen.getByTestId('date-time-picker-spinner').blur();
+    expect(screen.getByText('2020-04-01 13:34 to 2020-04-06 11:49')).toBeVisible();
+    userEvent.click(screen.getByText('Apply'));
+    expect(dateTimePickerProps.onApply).toHaveBeenCalled();
+  });
+
+  it('should render with a predefined single select date and 12 hours time format time', () => {
     const { i18n } = DateTimePicker.defaultProps;
     render(
       <DateTimePicker
@@ -482,6 +529,50 @@ describe('DateTimePickerV2', () => {
 
     expect(startTime).toHaveValue('05:03 PM');
     expect(screen.getByText('2020-04-01 05:03 PM')).toBeVisible();
+    expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
+  });
+
+  it('should render with a predefined single select date and 24 hours time format time', () => {
+    const { i18n } = DateTimePicker.defaultProps;
+    render(
+      <DateTimePicker
+        {...dateTimePickerProps}
+        useNewTimeSpinner
+        onApply={jest.fn()}
+        datePickerType="single"
+        dateTimeMask="YYYY-MM-DD HH:mm"
+        hasTimeInput
+        showRelativeOption={false}
+        defaultValue={{
+          timeRangeKind: PICKER_KINDS.SINGLE,
+          timeSingleValue: {
+            startDate: '2020-04-01',
+            startTime: '12:34',
+          },
+        }}
+      />
+    );
+
+    // default value is 2020-04-01 12:34
+    expect(screen.getByText('2020-04-01 12:34')).toBeVisible();
+
+    // first open the menu
+    userEvent.click(screen.getAllByText(/2020-04-01 12:34/i)[0]);
+
+    const startTime = screen.getByTestId('date-time-picker-input');
+    // open time picker
+    fireEvent.focus(screen.getByTestId('date-time-picker-input'));
+    // select hour
+    userEvent.click(
+      within(screen.getByTestId('date-time-picker-spinner-list-spinner-1')).getByText('23')
+    );
+    // select miniute
+    userEvent.click(
+      within(screen.getByTestId('date-time-picker-spinner-list-spinner-2')).getByText('03')
+    );
+
+    expect(startTime).toHaveValue('23:03');
+    expect(screen.getByText('2020-04-01 23:03')).toBeVisible();
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
   });
 
@@ -538,6 +629,38 @@ describe('DateTimePickerV2', () => {
 
     expect(screen.getByText('YYYY-MM-DD hh:mm A')).toBeVisible();
     expect(screen.queryByRole('dialog')).toBeNull();
+  });
+
+  it('should disable button when clear time picker input in single select', () => {
+    const { i18n } = DateTimePicker.defaultProps;
+    render(
+      <DateTimePicker
+        {...dateTimePickerProps}
+        useNewTimeSpinner
+        onApply={jest.fn()}
+        datePickerType="single"
+        dateTimeMask="YYYY-MM-DD hh:mm A"
+        hasTimeInput
+        showRelativeOption={false}
+        defaultValue={{
+          timeRangeKind: PICKER_KINDS.SINGLE,
+          timeSingleValue: {
+            startDate: '2020-04-01',
+            startTime: '12:34 AM',
+          },
+        }}
+      />
+    );
+
+    // default value is 2020-04-01 12:34 AM
+    expect(screen.getByText('2020-04-01 12:34 AM')).toBeVisible();
+
+    // first open the menu
+    userEvent.click(screen.getAllByText(/2020-04-01 12:34 AM/i)[0]);
+
+    userEvent.clear(screen.getByTestId('date-time-picker-input'));
+
+    expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
   });
 
   it('should go back to presets when cancel button is picked on absolute screen', () => {
@@ -841,7 +964,7 @@ describe('DateTimePickerV2', () => {
     );
     screen.getByTestId('date-time-picker-spinner').blur();
 
-    expect(screen.getByText('2020-04-01 13:34 to 2020-04-06 10:49')).toBeVisible();
+    expect(screen.getByText('2020-04-01 01:34 to 2020-04-06 10:49')).toBeVisible();
 
     userEvent.click(screen.getByText('Back'));
     userEvent.click(screen.getByText('Cancel'));
@@ -1042,6 +1165,8 @@ describe('DateTimePickerV2', () => {
     expect(screen.queryByText(i18nDefault.startTimeLabel)).not.toBeInTheDocument();
     expect(screen.queryByText(i18nDefault.endTimeLabel)).not.toBeInTheDocument();
     // click apply
+    expect(screen.getByText(i18nTest.applyBtnLabel)).toBeDisabled();
+    userEvent.click(screen.getAllByLabelText('Increment hours')[0]);
     fireEvent.click(screen.getByText(i18nTest.applyBtnLabel));
     expect(screen.getAllByTitle(new RegExp(`.*${i18nTest.toLabel}.*`))[0]).toBeInTheDocument();
 
@@ -1534,6 +1659,7 @@ describe('DateTimePickerV2', () => {
       <DateTimePicker
         {...dateTimePickerProps}
         useNewTimeSpinner
+        dateTimeMask="YYYY-MM-DD HH:mm A"
         id="picker-test"
         hasTimeInput
         defaultValue={{
@@ -1557,69 +1683,70 @@ describe('DateTimePickerV2', () => {
     const startTime = screen.getByTestId('date-time-picker-input-1');
     const endTime = screen.getByTestId('date-time-picker-input-2');
 
-    expect(timeRange).toHaveTextContent('2020-04-01 12:34 to 2020-04-01 11:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 12:34 PM to 2020-04-01 11:49 AM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
 
     // 2020-04-01 13:34 to 2020-04-01 11:49
     fireEvent.focus(startTime);
-    userEvent.type(
-      startTime,
-      '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}01:34 PM'
-    );
-    expect(startTime).toHaveValue('01:34 PM');
-    expect(endTime).toHaveValue('11:49 AM');
-    expect(timeRange).toHaveTextContent('2020-04-01 13:34 to 2020-04-01 11:49');
+    userEvent.type(startTime, '{backspace}{backspace}{backspace}{backspace}{backspace}13:34');
+    expect(startTime).toHaveValue('13:34');
+    expect(endTime).toHaveValue('11:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 13:34 PM to 2020-04-01 11:49 AM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
 
     // 2020-04-01 13:34 to 2020-04-01 12:49
     fireEvent.focus(endTime);
-    userEvent.type(
-      endTime,
-      '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}12:49 PM'
-    );
-    expect(startTime).toHaveValue('01:34 PM');
-    expect(endTime).toHaveValue('12:49 PM');
-    expect(timeRange).toHaveTextContent('2020-04-01 13:34 to 2020-04-01 12:49');
+    userEvent.type(endTime, '{backspace}{backspace}{backspace}{backspace}{backspace}12:49');
+    expect(startTime).toHaveValue('13:34');
+    expect(endTime).toHaveValue('12:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 13:34 PM to 2020-04-01 12:49 PM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
 
     // 2020-04-01 13:34 to 2020-04-01 13:49
-    userEvent.type(
-      endTime,
-      '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}01:49 PM'
-    );
+    userEvent.type(endTime, '{backspace}{backspace}{backspace}{backspace}{backspace}13:49');
     endTime.blur();
-    expect(startTime).toHaveValue('01:34 PM');
-    expect(endTime).toHaveValue('01:49 PM');
-    expect(timeRange).toHaveTextContent('2020-04-01 13:34 to 2020-04-01 13:49');
+    expect(startTime).toHaveValue('13:34');
+    expect(endTime).toHaveValue('13:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 13:34 PM to 2020-04-01 13:49 PM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     // 2020-04-01 13:50 to 2020-04-01 13:49
     fireEvent.focus(startTime);
-    userEvent.type(
-      startTime,
-      '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}01:50 PM'
-    );
-    expect(startTime).toHaveValue('01:50 PM');
-    expect(endTime).toHaveValue('01:49 PM');
-    expect(timeRange).toHaveTextContent('2020-04-01 13:50 to 2020-04-01 13:49');
+    userEvent.type(startTime, '{backspace}{backspace}{backspace}{backspace}{backspace}13:50');
+    expect(startTime).toHaveValue('13:50');
+    expect(endTime).toHaveValue('13:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 13:50 PM to 2020-04-01 13:49 PM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
 
-    // userEvent.type(screen.getByTestId('date-time-picker-input-1'), '{backspace}');
-    // userEvent.type(screen.getByTestId('date-time-picker-input-2'), '{backspace}');
-    // expect(screen.getByTestId('date-time-picker-input-1')).toHaveValue('Invalid Date');
-    // expect(screen.getByTestId('date-time-picker-input-2')).toHaveValue('Invalid Date');
-    // expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
-
-    userEvent.type(startTime, '{backspace}{backspace}{backspace}{backspace}{backspace}51 PM');
+    userEvent.type(startTime, '{backspace}{backspace}51');
 
     fireEvent.focus(endTime);
-    userEvent.type(
-      endTime,
-      '{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}{backspace}9999'
-    );
+    userEvent.type(endTime, '{backspace}{backspace}{backspace}{backspace}{backspace}9999');
     endTime.blur();
-    expect(startTime).toHaveValue('01:51 PM');
+    expect(startTime).toHaveValue('13:51');
     expect(endTime).toHaveValue('9999');
+    expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
+  });
+
+  it('should disable apply button when switching from presets to relative then to absolute with empty startDate and endDate', () => {
+    const { i18n } = DateTimePicker.defaultProps;
+    render(<DateTimePicker {...dateTimePickerProps} id="picker-test" />);
+    jest.runAllTimers();
+
+    userEvent.click(screen.getByTestId('date-time-picker__field'));
+    userEvent.click(screen.getByText('Custom Range'));
+    userEvent.click(screen.getByText('Absolute'));
+    expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
+  });
+
+  it('should disable apply button when switching from presets to relative then to absolute with empty startDate and endDate (new time spinner)', () => {
+    const { i18n } = DateTimePicker.defaultProps;
+    render(<DateTimePicker {...dateTimePickerProps} id="picker-test" useNewTimeSpinner />);
+    jest.runAllTimers();
+
+    userEvent.click(screen.getByTestId('date-time-picker__field'));
+    userEvent.click(screen.getByText('Custom Range'));
+    userEvent.click(screen.getByText('Absolute'));
     expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
   });
 
@@ -1698,6 +1825,7 @@ describe('DateTimePickerV2', () => {
       <DateTimePicker
         {...dateTimePickerProps}
         useNewTimeSpinner
+        dateTimeMask="YYYY-MM-DD hh:mm A"
         id="picker-test"
         testId="date-time-picker"
         hasTimeInput
@@ -1723,7 +1851,7 @@ describe('DateTimePickerV2', () => {
     const startTime = screen.getByTestId('date-time-picker-input-1');
     const endTime = screen.getByTestId('date-time-picker-input-2');
 
-    expect(timeRange).toHaveTextContent('2020-04-01 12:34 to 2020-04-06 11:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 12:34 PM to 2020-04-06 11:49 AM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     // 2020-04-01 13:34 to 2020-04-06 11:49
@@ -1735,7 +1863,7 @@ describe('DateTimePickerV2', () => {
 
     expect(startTime).toHaveValue('01:34 PM');
     expect(endTime).toHaveValue('11:49 AM');
-    expect(timeRange).toHaveTextContent('2020-04-01 13:34 to 2020-04-06 11:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 01:34 PM to 2020-04-06 11:49 AM');
 
     // 2020-04-01 13:34 to 2020-04-06 12:49
     fireEvent.focus(endTime);
@@ -1746,7 +1874,7 @@ describe('DateTimePickerV2', () => {
     endTime.blur();
     expect(startTime).toHaveValue('01:34 PM');
     expect(endTime).toHaveValue('12:49 PM');
-    expect(timeRange).toHaveTextContent('2020-04-01 13:34 to 2020-04-06 12:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 01:34 PM to 2020-04-06 12:49 PM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     // 2020-04-01 13:34 to 2020-04-06 13:49
@@ -1757,7 +1885,7 @@ describe('DateTimePickerV2', () => {
     );
     expect(startTime).toHaveValue('01:34 PM');
     expect(endTime).toHaveValue('01:49 PM');
-    expect(timeRange).toHaveTextContent('2020-04-01 13:34 to 2020-04-06 13:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 01:34 PM to 2020-04-06 01:49 PM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     // 2020-04-01 13:50 to 2020-04-06 13:49
@@ -1768,7 +1896,7 @@ describe('DateTimePickerV2', () => {
     );
     expect(screen.getByTestId('date-time-picker-input-1')).toHaveValue('01:50 PM');
     expect(screen.getByTestId('date-time-picker-input-2')).toHaveValue('01:49 PM');
-    expect(timeRange).toHaveTextContent('2020-04-01 13:50 to 2020-04-06 13:49');
+    expect(timeRange).toHaveTextContent('2020-04-01 01:50 PM to 2020-04-06 01:49 PM');
     expect(screen.getByText(i18n.applyBtnLabel)).toBeEnabled();
 
     fireEvent.focus(startTime);
@@ -1784,14 +1912,30 @@ describe('DateTimePickerV2', () => {
     expect(screen.getByText(i18n.applyBtnLabel)).toBeDisabled();
   });
 
+  it('should be able to generate default id', () => {
+    render(<DateTimePicker {...dateTimePickerProps} testId="date-time-picker" />);
+    expect(screen.getByTestId('date-time-picker').getAttribute('id')).toMatch(
+      /-iot--date-time-pickerv2__wrapper/
+    );
+  });
+
+  it('should be able to generate default id (new time spinner)', () => {
+    render(<DateTimePicker {...dateTimePickerProps} testId="date-time-picker" useNewTimeSpinner />);
+    expect(screen.getByTestId('date-time-picker').getAttribute('id')).toMatch(
+      /-iot--date-time-pickerv2__wrapper/
+    );
+  });
+
   it('should show invalid text when in invalid state', () => {
     render(<DateTimePicker {...dateTimePickerProps} i18n={i18n} invalid />);
     expect(screen.getByText(i18n.invalidText)).toBeInTheDocument();
     expect(
       screen
         .getByTestId(`${dateTimePickerProps.testId}-datepicker-flyout-button`)
-        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button-invalid`)
+        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button--invalid`)
     ).toBe(true);
+    userEvent.click(screen.getByText(i18n.invalidText));
+    expect(screen.queryByRole('dialog', { 'aria-labelledby': 'flyout-tooltip' })).toBeNull();
   });
 
   it('should show invalid text when in invalid state (new time spinner)', () => {
@@ -1800,8 +1944,10 @@ describe('DateTimePickerV2', () => {
     expect(
       screen
         .getByTestId(`${dateTimePickerProps.testId}-datepicker-flyout-button`)
-        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button-invalid`)
+        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button--invalid`)
     ).toBe(true);
+    userEvent.click(screen.getByText(i18n.invalidText));
+    expect(screen.queryByRole('dialog', { 'aria-labelledby': 'flyout-tooltip' })).toBeNull();
   });
 
   it('should disable menu button when disabled is set', () => {
@@ -1809,7 +1955,7 @@ describe('DateTimePickerV2', () => {
     expect(
       screen
         .getByTestId(`${dateTimePickerProps.testId}-datepicker-flyout-button`)
-        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button-disabled`)
+        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button--disabled`)
     ).toBe(true);
   });
 
@@ -1818,7 +1964,7 @@ describe('DateTimePickerV2', () => {
     expect(
       screen
         .getByTestId(`${dateTimePickerProps.testId}-datepicker-flyout-button`)
-        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button-disabled`)
+        .classList.contains(`${iotPrefix}--date-time-picker--trigger-button--disabled`)
     ).toBe(true);
   });
 });