+++ packages/react/src/components/DateTimePicker/DateTimePickerV2WithTimeSpinner.jsx
@@ -317,6 +317,12 @@ const DateTimePicker = ({
     }),
     [i18n]
   );
+
+  const is24hours = useMemo(() => {
+    const [, time] = dateTimeMask.split(' ');
+    const hoursMask = time.split(':')[0];
+    return hoursMask === 'HH';
+  }, [dateTimeMask]);
   const isSingleSelect = useMemo(() => datePickerType === 'single', [datePickerType]);
 
   // initialize the dayjs locale
@@ -342,8 +348,8 @@ const DateTimePicker = ({
     absoluteValue,
     setAbsoluteValue,
     resetAbsoluteValue,
-    format12hourTo24hour,
     isValid12HourTime,
+    isValid24HourTime,
   } = useAbsoluteDateTimeValue();
 
   const {
@@ -417,14 +423,13 @@ const DateTimePicker = ({
       } else if (customRangeKind === PICKER_KINDS.ABSOLUTE) {
         value.absolute = {
           ...absoluteValue,
-          startTime: hasTimeInput ? format12hourTo24hour(rangeStartTimeValue) : '00:00',
-          endTime: hasTimeInput ? format12hourTo24hour(rangeEndTimeValue) : '00:00',
+          startTime: hasTimeInput ? rangeStartTimeValue : '00:00',
+          endTime: hasTimeInput ? rangeEndTimeValue : '00:00',
         };
       } else {
         value.single = {
           ...singleDateValue,
-          startTime:
-            hasTimeInput && singleTimeValue !== '' ? format12hourTo24hour(singleTimeValue) : null,
+          startTime: hasTimeInput && singleTimeValue !== '' ? singleTimeValue : null,
         };
       }
       value.kind = customRangeKind;
@@ -582,9 +587,13 @@ const DateTimePicker = ({
           absolute.end = dayjs(`${absolute.endDate} ${absolute.endTime}`).valueOf();
         }
         absolute.startDate = dayjs(absolute.start).format('MM/DD/YYYY');
-        absolute.startTime = dayjs(absolute.start).format('hh:mm A');
+        absolute.startTime = is24hours
+          ? dayjs(absolute.start).format('HH:mm')
+          : dayjs(absolute.start).format('hh:mm A');
         absolute.endDate = dayjs(absolute.end).format('MM/DD/YYYY');
-        absolute.endTime = dayjs(absolute.end).format('hh:mm A');
+        absolute.endTime = is24hours
+          ? dayjs(absolute.end).format('HH:mm')
+          : dayjs(absolute.end).format('hh:mm A');
         setAbsoluteValue(absolute);
         setRangeStartTimeValue(absolute.startTime);
         setRangeEndTimeValue(absolute.endTime);
@@ -600,7 +609,11 @@ const DateTimePicker = ({
           single.start = dayjs(`${single.startDate} ${single.startTime}`).valueOf();
         }
         single.startDate = single.start ? dayjs(single.start).format('MM/DD/YYYY') : null;
-        single.startTime = single.start ? dayjs(single.start).format('hh:mm A') : null;
+        single.startTime = single.start
+          ? is24hours
+            ? dayjs(single.start).format('HH:mm')
+            : dayjs(single.start).format('hh:mm A')
+          : null;
         setSingleDateValue(single);
         setSingleTimeValue(single.startTime ?? '');
       }
@@ -659,10 +672,14 @@ const DateTimePicker = ({
   const disableAbsoluteApply =
     isCustomRange &&
     customRangeKind === PICKER_KINDS.ABSOLUTE &&
-    (invalidRangeStartTime || invalidRangeEndTime);
+    (invalidRangeStartTime ||
+      invalidRangeEndTime ||
+      (absoluteValue.startDate === '' && absoluteValue.endDate === ''));
 
   const disableSingleApply =
-    isCustomRange && customRangeKind === PICKER_KINDS.SINGLE && invalidRangeStartTime;
+    isCustomRange &&
+    customRangeKind === PICKER_KINDS.SINGLE &&
+    (invalidRangeStartTime || singleTimeValue === '');
 
   const disableApply = disableRelativeApply || disableAbsoluteApply || disableSingleApply;
 
@@ -798,17 +815,19 @@ const DateTimePicker = ({
     setRangeEndTimeValue(endState);
     setInvalidRangeStartTime(
       (absoluteValue && invalidStartDate(startState, endState, absoluteValue)) ||
-        !isValid12HourTime(startState)
+        (is24hours ? !isValid24HourTime(startState) : !isValid12HourTime(startState))
     );
     setInvalidRangeEndTime(
       (absoluteValue && invalidEndDate(startState, endState, absoluteValue)) ||
-        !isValid12HourTime(endState)
+        (is24hours ? !isValid24HourTime(endState) : !isValid12HourTime(endState))
     );
   };
 
   const handleSingleTimeValueChange = (startState) => {
     setSingleTimeValue(startState);
-    setInvalidRangeStartTime(!isValid12HourTime(startState));
+    setInvalidRangeStartTime(
+      is24hours ? !isValid24HourTime(startState) : !isValid12HourTime(startState)
+    );
   };
 
   return (
@@ -875,7 +894,7 @@ const DateTimePicker = ({
                   {humanValue}
                 </TooltipDefinition>
               ) : null}
-              {!isExpanded && isTooltipOpen ? (
+              {!isExpanded && isTooltipOpen && !isSingleSelect ? (
                 <Tooltip
                   open={isTooltipOpen}
                   showIcon={false}
@@ -898,11 +917,10 @@ const DateTimePicker = ({
             buttonProps={{
               tooltipPosition: 'top',
               tabIndex: -1,
-              className: invalidState
-                ? `${iotPrefix}--date-time-picker--trigger-button-invalid`
-                : disabled
-                ? `${iotPrefix}--date-time-picker--trigger-button-disabled`
-                : '',
+              className: classnames(`${iotPrefix}--date-time-picker--trigger-button`, {
+                [`${iotPrefix}--date-time-picker--trigger-button--invalid`]: invalid,
+                [`${iotPrefix}--date-time-picker--trigger-button--disabled`]: disabled,
+              }),
             }}
             hideTooltip
             iconDescription={mergedI18n.calendarLabel}
@@ -1164,6 +1182,7 @@ const DateTimePicker = ({
                           size="sm"
                           testId={testId}
                           style={{ zIndex: (style.zIndex ?? 0) + 6000 }}
+                          is24hours={is24hours}
                         />
                       ) : (
                         <div className={`${iotPrefix}--date-time-picker__no-formgroup`} />
@@ -1175,17 +1194,17 @@ const DateTimePicker = ({
             </div>
           </FlyoutMenu>
         </div>
-        {invalidState && !hasIconOnly ? (
-          <p
-            className={classnames(
-              `${prefix}--form__helper-text`,
-              `${iotPrefix}--date-time-picker__helper-text--invalid`
-            )}
-          >
-            {mergedI18n.invalidText}
-          </p>
-        ) : null}
       </div>
+      {invalidState && !hasIconOnly ? (
+        <p
+          className={classnames(
+            `${prefix}--form__helper-text`,
+            `${iotPrefix}--date-time-picker__helper-text--invalid`
+          )}
+        >
+          {mergedI18n.invalidText}
+        </p>
+      ) : null}
     </div>
   );
 };