+++ packages/react/src/components/Table/Table.jsx
@@ -239,6 +239,8 @@ const propTypes = {
       rowEditBarButtons: PropTypes.node,
       /** extra actions that can appear in an overflow menu in the toolbar (same menu as toggle aggregations) */
       toolbarActions: TableToolbarActionsPropType,
+      /** force hide Clear all filters button in toolbar */
+      hideClearAllFiltersButton: PropTypes.bool,
     }),
     table: PropTypes.shape({
       isSelectAllSelected: PropTypes.bool,
@@ -317,6 +319,8 @@ const propTypes = {
       onToggleAggregations: PropTypes.func,
       /** fired when clicking a 'toolbarAction' in the table toolbar */
       onApplyToolbarAction: PropTypes.func,
+      /** fired when focus/blur on search bar in the table toolbar */
+      onSearchExpand: PropTypes.func,
     }),
     /** table wide actions */
     table: PropTypes.shape({
@@ -415,6 +419,7 @@ export const defaultProps = (baseProps) => ({
       advancedFilterFlyoutOpen: false,
       batchActions: [],
       search: {},
+      hideClearAllFiltersButton: false,
     },
     table: {
       expandedIds: [],
@@ -875,7 +880,8 @@ const Table = (props) => {
                 'onChangeAdvancedFilter',
                 'onRemoveAdvancedFilter',
                 'onToggleAdvancedFilter',
-                'onApplyToolbarAction'
+                'onApplyToolbarAction',
+                'onSearchExpand'
               ),
               onToggleAggregations,
               onApplySearch: (value) => {
@@ -921,6 +927,7 @@ const Table = (props) => {
                 'advancedFilterFlyoutOpen',
                 'toolbarActions'
               ),
+              hideClearAllFiltersButton: view.toolbar.hideClearAllFiltersButton,
             }}
             data={data}
             // TODO: remove id in V3.