+++ packages/react/src/components/Table/Table.story.helpers.jsx
@@ -117,6 +117,7 @@ export const getTableActions = () => ({
     // can be readded in V3.
     // onToggleAggregations: action('onToggleAggregations'),
     onApplyToolbarAction: action('onApplyToolbarAction'),
+    onSearchExpand: action('onSearchExpand'),
   },
   table: {
     onRowClicked: action('onRowClicked'),
@@ -1045,6 +1046,13 @@ export const getTableKnobs = ({ knobsToCreate, getDefaultValue, useGroups = fals
       ? boolean(
           'Enable advanced filters (options.hasAdvancedFilter) ☢️',
           getDefaultValue('hasAdvancedFilter'),
+          SORT_FILTER_GROUP
+        )
+      : null,
+    hideClearAllFiltersButton: shouldCreate('hideClearAllFiltersButton')
+      ? boolean(
+          'Hide clear all filters button (view.toolbar.hideClearAllFiltersButton)',
+          getDefaultValue('hideClearAllFiltersButton'),
           SORT_FILTER_GROUP
         )
       : null,